Move code block extraction to utilities module
